# openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: BBLBB API
  version: 1.0.0
  description: |
    BBLBB 稳定业务 API。字段、错误、权限和安全语义以本文件及其引用的冻结文档为准。
    生产实现必须保持 SQLite、MySQL 8 和 MariaDB 10.11 的契约一致性。
  license:
    name: Proprietary
servers:
- url: https://{host}
  variables:
    host:
      default: forum.example.com
security: []
...
paths:
...
  "/api/v1/auth/passkeys":
    get:
      tags:
      - Auth
      operationId: listPasskeys
      security:
      - sessionCookie: []
      x-permission: passkey.read_own
      summary: List the current user's registered passkeys
      x-csrf: false
      responses:
        '200':
          description: Registered passkeys (credential secrets never returned)
          headers:
            Cache-Control:
              description: private, no-store
              schema:
                type: string
                example: private, no-store
            X-Request-ID:
              "$ref": "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/PasskeyListResponse"
        '401':
          "$ref": "#/components/responses/AuthenticationRequired"
        '429':
          "$ref": "#/components/responses/RateLimited"
    post:
      tags:
      - Auth
      operationId: beginPasskeyRegistration
      security:
      - sessionCookie: []
      x-permission: passkey.enroll
      summary: Begin passkey registration (returns WebAuthn creation options)
      x-csrf: true
      parameters:
      - "$ref": "#/components/parameters/IdempotencyKey"
      responses:
        '200':
          description: WebAuthn PublicKeyCredentialCreationOptions (passkey mode)
          headers:
            Cache-Control:
              description: private, no-store
              schema:
                type: string
                example: private, no-store
            X-Request-ID:
              "$ref": "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/PasskeyRegistrationOptions"
        '401':
          "$ref": "#/components/responses/AuthenticationRequired"
        '403':
          "$ref": "#/components/responses/CsrfFailed"
        '429':
          "$ref": "#/components/responses/RateLimited"
  "/api/v1/auth/passkeys/confirm":
    post:
      tags:
      - Auth
      operationId: confirmPasskeyRegistration
      security:
      - sessionCookie: []
      x-permission: passkey.enroll
      summary: Confirm passkey registration with the browser credential response
      x-csrf: true
      parameters:
      - "$ref": "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/PasskeyConfirmRequest"
      responses:
        '200':
          "$ref": "#/components/responses/GenericSuccess"
        '400':
          "$ref": "#/components/responses/ProblemResponse"
        '401':
          "$ref": "#/components/responses/AuthenticationRequired"
        '403':
          "$ref": "#/components/responses/CsrfFailed"
        '429':
          "$ref": "#/components/responses/RateLimited"
  "/api/v1/auth/passkeys/{id}":
    delete:
      tags:
      - Auth
      operationId: revokePasskey
      security:
      - sessionCookie: []
      x-permission: passkey.revoke
      summary: Revoke one of the current user's passkeys (recent auth required)
      x-csrf: true
      parameters:
      - name: id
        in: path
        required: true
        description: Passkey 记录 ID（GET /auth/passkeys 返回的内部 id）
        schema:
          type: string
      - "$ref": "#/components/parameters/IdempotencyKey"
      responses:
        '200':
          "$ref": "#/components/responses/GenericSuccess"
        '401':
          "$ref": "#/components/responses/AuthenticationRequired"
        '403':
          "$ref": "#/components/responses/StepUpRequired"
        '404':
          "$ref": "#/components/responses/NotFound"
        '429':
          "$ref": "#/components/responses/RateLimited"
  "/api/v1/auth/login/mfa/passkey/options":
    post:
      tags:
      - Auth
      operationId: loginMfaPasskeyOptions
      security: *1
      x-permission: public
      summary: Get WebAuthn request options for the passkey step of a two-step login
      description: 用第一步 /login 返回的一次性 challenge 换取 PublicKeyCredentialRequestOptions（仅当该账号注册过 Passkey 时可用）。
      x-csrf: true
      x-csrf-context: preauth
      parameters:
      - "$ref": "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/PasskeyLoginOptionsRequest"
      responses:
        '200':
          description: WebAuthn PublicKeyCredentialRequestOptions（allowCredentials 限定该账号凭据）
          headers:
            Cache-Control:
              description: private, no-store
              schema:
                type: string
                example: private, no-store
            X-Request-ID:
              "$ref": "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/PasskeyRequestOptions"
        '400':
          "$ref": "#/components/responses/ProblemResponse"
        '401':
          "$ref": "#/components/responses/AuthenticationRequired"
        '403':
          "$ref": "#/components/responses/CsrfFailed"
        '422':
          "$ref": "#/components/responses/ValidationError"
        '429':
          "$ref": "#/components/responses/RateLimited"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-bblbb_session
      description: Browser session cookie. State-changing requests also require X-CSRF-Token.
    bearerToken:
      type: http
      scheme: bearer
      bearerFormat: opaque
  headers:
    RequestId:
      description: UUID v7 request correlation ID.
      schema:
        type: string
        format: uuid
...
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 200
...
  responses:
    AuthenticationRequired:
      description: Authentication is required
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/Problem"
...
    CsrfFailed:
      description: CSRF validation failed
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/Problem"
    StepUpRequired:
      description: Recent authentication required (M02-MFA-07 step-up); re-authenticate via /api/v1/auth/re-auth
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/Problem"
    NotFound:
      description: Resource is missing or intentionally hidden
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/Problem"
...
    ValidationError:
      description: Request validation failed
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/Problem"
...
    RateLimited:
      description: Request rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
            minimum: 1
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/Problem"
...
    GenericSuccess:
      description: Successful operation
      headers:
        Cache-Control:
          description: private, no-store
          schema:
            type: string
            example: private, no-store
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/GenericSuccess"
    ProblemResponse:
      description: Request failed
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/Problem"
  schemas:
...
    GenericSuccess:
      type: object
      additionalProperties: true
...
    PasskeyLoginOptionsRequest:
      type: object
      additionalProperties: false
      required:
      - challenge_token
      properties:
        challenge_token:
          type: string
          description: 第一步 /login 返回的一次性 MFA challenge
    PasskeyRegistrationOptions:
      type: object
      description: WebAuthn PublicKeyCredentialCreationOptions（透传给 navigator.credentials.create()）
      required:
      - publicKey
      properties:
        publicKey:
          type: object
          required:
          - rp
          - user
          - challenge
          - pubKeyCredParams
          properties:
            rp:
              type: object
              properties:
                id:
                  type: string
                name:
                  type: string
            user:
              type: object
              properties:
                id:
                  type: string
                  description: base64url 编码的服务端用户标识
                name:
                  type: string
                displayName:
                  type: string
            challenge:
              type: string
              description: base64url 编码的服务端 challenge
            pubKeyCredParams:
              type: array
              items:
                type: object
                properties:
                  type:
                    type: string
                  alg:
                    type: integer
            timeout:
              type: integer
            excludeCredentials:
              type: array
              description: 该账号已注册凭据（防重复注册同一认证器）
              items:
                type: object
                properties:
                  type:
                    type: string
                  id:
                    type: string
                  transports:
                    type: array
                    items:
                      type: string
            authenticatorSelection:
              type: object
              properties:
                residentKey:
                  type: string
                requireResidentKey:
                  type: boolean
                userVerification:
                  type: string
            attestation:
              type: string
          additionalProperties: true
      additionalProperties: false
    PasskeyConfirmRequest:
      type: object
      additionalProperties: false
      required:
      - credential
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 64
          description: 用户可读标签（如「MacBook 指纹」）；缺省为「Passkey」
        credential:
          "$ref": "#/components/schemas/PasskeyRegistration"
    PasskeyRegistration:
      type: object
      description: navigator.credentials.create() 返回的 PublicKeyCredential（浏览器原始 JSON）
      required:
      - id
      - rawId
      - type
      - response
      properties:
        id:
          type: string
        rawId:
          type: string
        type:
          type: string
          enum:
          - public-key
        response:
          type: object
          required:
          - clientDataJSON
          - attestationObject
          properties:
            clientDataJSON:
              type: string
            attestationObject:
              type: string
            transports:
              type: array
              items:
                type: string
        clientExtensionResults:
          type: object
          additionalProperties: true
        authenticatorAttachment:
          type: string
      additionalProperties: true
    PasskeyRequestOptions:
      type: object
      description: WebAuthn PublicKeyCredentialRequestOptions（透传给 navigator.credentials.get()）
      required:
      - publicKey
      properties:
        publicKey:
          type: object
          required:
          - challenge
          properties:
            rpId:
              type: string
            challenge:
              type: string
              description: base64url 编码的服务端 challenge
            timeout:
              type: integer
            allowCredentials:
              type: array
              description: 该账号已注册的 Passkey（服务端按登录 challenge 绑定生成）
              items:
                type: object
                properties:
                  type:
                    type: string
                  id:
                    type: string
                  transports:
                    type: array
                    items:
                      type: string
            userVerification:
              type: string
          additionalProperties: true
      additionalProperties: false
    PasskeyCredentialInfo:
      type: object
      additionalProperties: false
      required:
      - id
      - name
      - created_at
      properties:
        id:
          type: string
          description: Passkey 记录内部 ID（DELETE /auth/passkeys/{id} 使用）
        name:
          type: string
        aaguid:
          type: string
          nullable: true
          description: 认证器型号 GUID（未知为 null）
        backup_eligible:
          type: boolean
        backed_up:
          type: boolean
        created_at:
          type: integer
          format: int64
          description: 注册时间（Unix 毫秒）
        last_used_at:
          type: integer
          format: int64
          nullable: true
          description: 最近一次断言成功时间（Unix 毫秒；尚未使用为 null）
    PasskeyListResponse:
      type: object
      additionalProperties: false
      required:
      - passkeys
      properties:
        passkeys:
          type: array
          items:
            "$ref": "#/components/schemas/PasskeyCredentialInfo"
...
    Problem:
      type: object
      required:
      - type
      - title
      - status
      - code
      - detail
      - request_id
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        code:
          type: string
          enum:
          - invalid_request
          - visibility_level_exceeds_author
          - invalid_url
          - idempotency_conflict
          - version_conflict
          - unauthorized
          - invalid_credentials
          - mfa_code_invalid
          - mfa_challenge_invalid
          - mfa_confirm_invalid
          - mfa_enrollment_invalid
          - passkey_not_configured
          - passkey_unavailable
          - passkey_challenge_invalid
          - passkey_registration_invalid
          - passkey_limit_reached
          - reauth_password_invalid
          - verification_token_invalid
          - reset_token_invalid
          - validation_failed
          - invalid_current_password
          - forbidden
          - step_up_required
          - not_found
          - csrf_failed
          - origin_not_allowed
          - host_not_allowed
          - rate_limited
          - crawler_denied
          - challenge_required
          - temporarily_banned
          - feature_disabled
          - insufficient_funds
          - daily_limit_exceeded
          - checkout_interaction_invalid
          - checkout_user_mismatch
          - checkout_intent_expired
          - checkout_intent_consumed
          - offer_version_changed
          - refund_not_allowed
          - product_unavailable
          - shop_purchase_limit_exceeded
          - shop_stock_exhausted
          - entitlement_not_usable
          - presentation_slot_conflict
          - activity_already_claimed
          - activity_not_eligible
          - invalid_price_coin
          - post_not_paid
          - price_not_configured
          - self_reaction
          - cannot_follow_self
          - cannot_message_self
          - download_url_unavailable
          - ai_consent_required
          - ai_budget_exceeded
          - ai_suggestion_stale
          - invalid_storage_request
          - storage_partial_upload
          - storage_forbidden
          - storage_auth_failed
          - storage_rate_limited
          - quota_exceeded
          - storage_conflict
          - storage_state_error
          - storage_verification_failed
          - storage_network_error
          - storage_upstream_error
          - theme_invalid
          - theme_incompatible
          - theme_not_found
          - theme_conflict
          - plugin_invalid
          - plugin_incompatible
          - plugin_not_found
          - plugin_conflict
          - marketplace_disabled
          - marketplace_invalid_client
          - refund_exceeds_purchase
          - merchant_balance_insufficient
          - webhook_invalid_signature
          - bad_request
          - conflict
          - video_insecure_scheme
          - video_invalid_url
          - video_host_invalid
          - video_port_not_allowed
          - video_private_ip
          - video_signed_url
          - video_userinfo_not_allowed
          - video_fragment_not_allowed
          - video_unsupported_type
          - video_not_video_page
          - video_invalid
          - video_mime_mismatch
          - video_no_embed_permission
          - video_takedown
          - video_provider_disabled
          - video_provider_host_not_allowed
          - video_provider_ratelimited
          - video_provider_unavailable
          - video_policy_changed
          - video_policy_version_conflict
          - video_poster_attachment_invalid
          - video_resolution_expired
          - video_embed_not_found
          - video_embed_referenced
          - video_target_conflict
          - video_target_forbidden
          - video_target_not_found
          - video_version_conflict
          - video_egress_http_error
          - video_egress_private_ip
          - video_egress_timeout
          - video_egress_too_large
          - video_egress_too_many_redirects
          - video_egress_unavailable
          - video_hls_invalid
          - video_hls_depth_exceeded
          - video_hls_segment_count_exceeded
          - video_hls_duration_exceeded
          - video_hls_cross_origin_segment
          - video_hls_key_not_allowed
          - video_hls_map_not_allowed
          - video_hls_signed_uri
          - internal_error
        detail:
          type: string
        instance:
          type: string
        request_id:
          type: string
          format: uuid
        errors:
          type: array
          items:
            type: object
            required:
            - code
            - message_key
            properties:
              field:
                type: string
              code:
                type: string
              message_key:
                type: string
...